requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
airtable-python-wrapper==0.15.3
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
asyncio==4.0.0
attrs==25.3.0
azure-core==1.35.1
azure-identity==1.25.0
azure-storage-blob==12.26.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.2
deprecation==2.1.0
exceptiongroup==1.3.0
eyeD3==0.9.8
fastapi==0.118.0
filetype==1.2.0
Flask==3.1.2
frozenlist==1.7.0
google-api-core==2.25.1
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-storage==3.4.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
msal==1.34.0
msal-extensions==1.3.1
multidict==6.6.4
mutagen==1.47.0
numpy==2.2.6
oauth2client==4.1.3
oauthlib==3.3.1
onedrivesdk_fork==1.1.14
packaging==25.0
pandas==2.3.3
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.1.0
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.5
Pyrogram==2.0.106
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==22.5
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
schedule==1.2.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.48.0
Telethon==1.41.2
TgCrypto==1.2.5
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
Werkzeug==3.1.3
yarl==1.20.1
